@@ -5,7 +5,8 @@ import zarr
 from scipy import sparse
 
 from anndata import AnnData
-from anndata.experimental import sparse_dataset, write_elem
+from anndata._core.sparse_dataset import sparse_dataset
+from anndata._io.specs import write_elem
 
 
 def make_alternating_mask(n):
