@@ -1,4 +1,4 @@
-#!python3
+#!/usr/bin/env python3
 from __future__ import annotations
 
 import argparse
