@@ -26,26 +26,21 @@ import numpy as np
 import scipy.sparse as ss
 from scipy.sparse import _sparsetools
 
-from anndata._core.index import _fix_slice_bounds
-from anndata.compat import H5Group, ZarrArray, ZarrGroup
-
-from ..compat import SpArray, _read_attr
-
-try:
-    # Not really important, just for IDEs to be more helpful
-    from scipy.sparse._compressed import _cs_matrix
-except ImportError:
-    from scipy.sparse import spmatrix as _cs_matrix
-
-
-from .index import _subset, unpack_index
+from .. import abc
+from .._settings import settings
+from ..compat import H5Group, SpArray, ZarrArray, ZarrGroup, _read_attr
+from .index import _fix_slice_bounds, _subset, unpack_index
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
     from typing import Literal
 
+    from scipy.sparse._compressed import _cs_matrix
+
     from .._types import GroupStorageType
     from .index import Index
+else:
+    from scipy.sparse import spmatrix as _cs_matrix
 
 
 class BackedFormat(NamedTuple):
@@ -234,6 +229,8 @@ class backed_csc_matrix(BackedSparseMatrix, ss.csc_matrix):
 FORMATS = [
     BackedFormat("csr", backed_csr_matrix, ss.csr_matrix),
     BackedFormat("csc", backed_csc_matrix, ss.csc_matrix),
+    BackedFormat("csr", backed_csr_matrix, ss.csr_array),
+    BackedFormat("csc", backed_csc_matrix, ss.csc_array),
 ]
 
 
@@ -346,24 +343,18 @@ def _get_group_format(group: GroupStorageType) -> str:
 def is_sparse_indexing_overridden(format: Literal["csr", "csc"], row, col):
     major_indexer, minor_indexer = (row, col) if format == "csr" else (col, row)
     return isinstance(minor_indexer, slice) and (
-        (isinstance(major_indexer, (int, np.integer)))
+        (isinstance(major_indexer, int | np.integer))
         or (isinstance(major_indexer, slice))
         or (isinstance(major_indexer, np.ndarray) and major_indexer.ndim == 1)
     )
 
 
-class BaseCompressedSparseDataset(ABC):
-    """Analogous to :class:`h5py.Dataset <h5py:Dataset>` or `zarr.Array`, but for sparse matrices."""
-
-    format: Literal["csr", "csc"]
+class BaseCompressedSparseDataset(abc._AbstractCSDataset, ABC):
     _group: GroupStorageType
 
     def __init__(self, group: GroupStorageType):
         type(self)._check_group_format(group)
         self._group = group
-
-    shape: tuple[int, int]
-    """Shape of the matrix."""
 
     @property
     def group(self) -> GroupStorageType:
@@ -378,6 +369,7 @@ class BaseCompressedSparseDataset(ABC):
 
     @property
     def backend(self) -> Literal["zarr", "hdf5"]:
+        """Which file type is used on-disk."""
         if isinstance(self.group, ZarrGroup):
             return "zarr"
         elif isinstance(self.group, H5Group):
@@ -387,6 +379,7 @@ class BaseCompressedSparseDataset(ABC):
 
     @property
     def dtype(self) -> np.dtype:
+        """The :class:`numpy.dtype` of the `data` attribute of the sparse matrix."""
         return self.group["data"].dtype
 
     @classmethod
@@ -395,43 +388,26 @@ class BaseCompressedSparseDataset(ABC):
         assert group_format == cls.format
 
     @property
-    def format_str(self) -> Literal["csr", "csc"]:
-        """DEPRECATED Use .format instead."""
-        warnings.warn(
-            "The attribute .format_str is deprecated and will be removed in the anndata 0.11.0. "
-            "Please use .format instead.",
-            FutureWarning,
-        )
-        return self.format
-
-    @property
-    def name(self) -> str:
+    def _name(self) -> str:
+        """Name of the group."""
         return self.group.name
 
     @property
     def shape(self) -> tuple[int, int]:
+        """Shape of the matrix read off disk."""
         shape = _read_attr(self.group.attrs, "shape", None)
         if shape is None:
             # TODO warn
             shape = self.group.attrs.get("h5sparse_shape")
         return tuple(map(int, shape))
 
-    @property
-    def value(self) -> ss.csr_matrix | ss.csc_matrix:
-        """DEPRECATED Use .to_memory() instead."""
-        warnings.warn(
-            "The .value attribute is deprecated and will be removed in the anndata 0.11.0. "
-            "Please use .to_memory() instead.",
-            FutureWarning,
-        )
-        return self.to_memory()
-
     def __repr__(self) -> str:
-        return f"{type(self).__name__}: backend {self.backend}, shape {self.shape}, data_dtype {self.dtype}"
+        name = type(self).__name__.removeprefix("_")
+        return f"{name}: backend {self.backend}, shape {self.shape}, data_dtype {self.dtype}"
 
     def __getitem__(
         self, index: Index | tuple[()]
-    ) -> float | ss.csr_matrix | ss.csc_matrix:
+    ) -> float | ss.csr_matrix | ss.csc_matrix | SpArray:
         indices = self._normalize_index(index)
         row, col = indices
         mtx = self._to_backed()
@@ -458,8 +434,15 @@ class BaseCompressedSparseDataset(ABC):
 
         # If indexing is array x array it returns a backed_sparse_matrix
         # Not sure what the performance is on that operation
-        if isinstance(sub, BackedSparseMatrix):
-            return get_memory_class(self.format)(sub)
+        # Also need to check if memory format is not matrix
+        mtx_fmt = get_memory_class(
+            self.format, use_sparray_in_io=settings.use_sparse_array_on_read
+        )
+        must_convert_to_array = issubclass(mtx_fmt, SpArray) and not isinstance(
+            sub, SpArray
+        )
+        if isinstance(sub, BackedSparseMatrix) or must_convert_to_array:
+            return mtx_fmt(sub)
         else:
             return sub
 
@@ -483,7 +466,25 @@ class BaseCompressedSparseDataset(ABC):
         mock_matrix[row, col] = value
 
     # TODO: split to other classes?
-    def append(self, sparse_matrix: _cs_matrix | SpArray) -> None:
+    def append(self, sparse_matrix: ss.csr_matrix | ss.csc_matrix | SpArray) -> None:
+        """Append an in-memory or on-disk sparse matrix to the current object's store.
+
+        Parameters
+        ----------
+        sparse_matrix
+            The matrix to append.
+
+        Raises
+        ------
+        NotImplementedError
+            If the matrix to append is not one of :class:`~scipy.sparse.csr_array`, :class:`~scipy.sparse.csc_array`, :class:`~scipy.sparse.csr_matrix`, or :class:`~scipy.sparse.csc_matrix`.
+        ValueError
+            If both the on-disk and to-append matrices are not of the same format i.e., `csr` or `csc`.
+        OverflowError
+            If the underlying data store has a 32 bit indptr, and the new matrix is too large to fit in it i.e., would cause a 64 bit `indptr` to be written.
+        AssertionError
+            If the on-disk data does not have `csc` or `csr` format.
+        """
         # Prep variables
         shape = self.shape
         if isinstance(sparse_matrix, BaseCompressedSparseDataset):
@@ -546,7 +547,7 @@ class BaseCompressedSparseDataset(ABC):
         )
         # Clear cached property
         if hasattr(self, "indptr"):
-            del self.indptr
+            del self._indptr
 
         # indices
         indices = self.group["indices"]
@@ -555,7 +556,7 @@ class BaseCompressedSparseDataset(ABC):
         indices[orig_data_size:] = sparse_matrix.indices
 
     @cached_property
-    def indptr(self) -> np.ndarray:
+    def _indptr(self) -> np.ndarray:
         """\
         Other than `data` and `indices`, this is only as long as the major axis
 
@@ -569,39 +570,29 @@ class BaseCompressedSparseDataset(ABC):
         mtx = format_class(self.shape, dtype=self.dtype)
         mtx.data = self.group["data"]
         mtx.indices = self.group["indices"]
-        mtx.indptr = self.indptr
+        mtx.indptr = self._indptr
         return mtx
 
-    def to_memory(self) -> ss.csr_matrix | ss.csc_matrix:
-        format_class = get_memory_class(self.format)
+    def to_memory(self) -> ss.csr_matrix | ss.csc_matrix | SpArray:
+        format_class = get_memory_class(
+            self.format, use_sparray_in_io=settings.use_sparse_array_on_read
+        )
         mtx = format_class(self.shape, dtype=self.dtype)
         mtx.data = self.group["data"][...]
         mtx.indices = self.group["indices"][...]
-        mtx.indptr = self.indptr
+        mtx.indptr = self._indptr
         return mtx
 
 
-_sparse_dataset_doc = """\
-    On disk {format} sparse matrix.
-
-    Parameters
-    ----------
-    group
-        The backing group store.
-"""
+class _CSRDataset(BaseCompressedSparseDataset, abc.CSRDataset):
+    """Internal concrete version of :class:`anndata.abc.CSRDataset`."""
 
 
-class CSRDataset(BaseCompressedSparseDataset):
-    __doc__ = _sparse_dataset_doc.format(format="CSR")
-    format = "csr"
+class _CSCDataset(BaseCompressedSparseDataset, abc.CSCDataset):
+    """Internal concrete version of :class:`anndata.abc.CSRDataset`."""
 
 
-class CSCDataset(BaseCompressedSparseDataset):
-    __doc__ = _sparse_dataset_doc.format(format="CSC")
-    format = "csc"
-
-
-def sparse_dataset(group: GroupStorageType) -> CSRDataset | CSCDataset:
+def sparse_dataset(group: GroupStorageType) -> abc.CSRDataset | abc.CSCDataset:
     """Generates a backed mode-compatible sparse dataset class.
 
     Parameters
@@ -620,7 +611,8 @@ def sparse_dataset(group: GroupStorageType) -> CSRDataset | CSCDataset:
 
     >>> import scanpy as sc
     >>> import h5py
-    >>> from anndata.experimental import sparse_dataset, read_elem
+    >>> from anndata.io import sparse_dataset
+    >>> from anndata.io import read_elem
     >>> sc.datasets.pbmc68k_reduced().raw.to_adata().write_h5ad("pbmc.h5ad")
 
     Initialize a sparse dataset from storage
@@ -653,38 +645,12 @@ def sparse_dataset(group: GroupStorageType) -> CSRDataset | CSCDataset:
     """
     encoding_type = _get_group_format(group)
     if encoding_type == "csr":
-        return CSRDataset(group)
+        return _CSRDataset(group)
     elif encoding_type == "csc":
-        return CSCDataset(group)
+        return _CSCDataset(group)
+    raise ValueError(f"Unknown encoding type {encoding_type}")
 
 
 @_subset.register(BaseCompressedSparseDataset)
 def subset_sparsedataset(d, subset_idx):
     return d[subset_idx]
-
-
-## Backwards compat
-
-_sparsedataset_depr_msg = """\
-SparseDataset is deprecated and will be removed in late 2024. It has been replaced by the public classes CSRDataset and CSCDataset.
-
-For instance checks, use `isinstance(X, (anndata.experimental.CSRDataset, anndata.experimental.CSCDataset))` instead.
-
-For creation, use `anndata.experimental.sparse_dataset(X)` instead.
-"""
-
-
-class SparseDataset(ABC):
-    """DEPRECATED.
-
-    Use CSRDataset, CSCDataset, and sparse_dataset from anndata.experimental instead.
-    """
-
-    def __new__(cls, group):
-        warnings.warn(FutureWarning(_sparsedataset_depr_msg), stacklevel=2)
-        return sparse_dataset(group)
-
-    @classmethod
-    def __subclasshook__(cls, C):
-        warnings.warn(FutureWarning(_sparsedataset_depr_msg), stacklevel=3)
-        return issubclass(C, (CSRDataset, CSCDataset))
