@@ -1,38 +1,52 @@
 from __future__ import annotations
 
-from anndata._core.sparse_dataset import CSCDataset, CSRDataset, sparse_dataset
-from anndata._io.specs import IOSpec, read_elem, read_elem_as_dask, write_elem
+from types import MappingProxyType
+from typing import TYPE_CHECKING
 
-from .._types import InMemoryElem as _InMemoryElem
+from .._io.specs import IOSpec, read_elem_as_dask
 from .._types import Read, ReadCallback, StorageType, Write, WriteCallback
-from .._types import RWAble as _RWAble
+from ..utils import module_get_attr_redirect
 from ._dispatch_io import read_dispatched, write_dispatched
 from .merge import concat_on_disk
 from .multi_files import AnnCollection
 from .pytorch import AnnLoader
 
-# Sphinx can’t find data docstrings when objects are re-exported
-InMemoryElem = _InMemoryElem
-"""An in-memory element that can be read and written, including an :class:`anndata.AnnData` objects."""
-RWAble = _RWAble
-"""A serializable object, excluding :class:`anndata.AnnData` objects i.e., something that can be stored in `uns` or `obsm`."""
+if TYPE_CHECKING:
+    from typing import Any
+
+
+# Map old name in `anndata.experimental` to new name in `anndata`
+_DEPRECATED = MappingProxyType(
+    dict(
+        (kv if isinstance(kv, tuple) else (kv, kv))
+        for kv in (
+            ("CSRDataset", "abc.CSRDataset"),
+            ("CSCDataset", "abc.CSCDataset"),
+            ("sparse_dataset", "io.sparse_dataset"),
+            ("read_elem", "io.read_elem"),
+            ("write_elem", "io.write_elem"),
+            ("RWAble", "typing.AxisStorable"),
+            ("InMemoryElem", "typing.RWAble"),
+        )
+    )
+)
+
+
+def __getattr__(attr_name: str) -> Any:
+    return module_get_attr_redirect(
+        attr_name, deprecated_mapping=_DEPRECATED, old_module_path="experimental"
+    )
+
 
 __all__ = [
     "AnnCollection",
     "AnnLoader",
-    "read_elem",
-    "write_elem",
     "read_elem_as_dask",
     "read_dispatched",
     "write_dispatched",
     "IOSpec",
     "concat_on_disk",
-    "sparse_dataset",
-    "CSRDataset",
-    "CSCDataset",
-    "InMemoryElem",
     "Read",
-    "RWAble",
     "Write",
     "ReadCallback",
     "WriteCallback",
