@@ -32,16 +32,24 @@ def pytest_configure(config: pytest.Config) -> None:
 
 
 @pytest.fixture(autouse=True)
-def _suppress_env_for_doctests(request: pytest.FixtureRequest) -> None:
+def _anndata_test_env(request: pytest.FixtureRequest) -> None:
+    import anndata
+
     if isinstance(request.node, pytest.DoctestItem):
         request.getfixturevalue("_doctest_env")
+
+    anndata.settings.reset(anndata.settings._registered_options.keys())
 
 
 @pytest.fixture
 def _doctest_env(
     request: pytest.FixtureRequest, cache: pytest.Cache, tmp_path: Path
 ) -> Generator[None, None, None]:
-    from scanpy import settings
+    with warnings.catch_warnings():
+        warnings.filterwarnings(
+            "ignore", message=r"Importing read_.* from `anndata` is deprecated"
+        )
+        from scanpy import settings
 
     from anndata.compat import chdir
     from anndata.utils import import_name
