@@ -10,8 +10,8 @@ from scipy import sparse
 
 from anndata import AnnData, concat
 from anndata._core.merge import _resolve_axis
-from anndata.experimental import read_elem, write_elem
 from anndata.experimental.merge import as_group, concat_on_disk
+from anndata.io import read_elem, write_elem
 from anndata.tests.helpers import (
     assert_equal,
     gen_adata,
