@@ -8,12 +8,7 @@ from scipy import sparse
 
 import anndata as ad
 from anndata.compat import SpArray
-from anndata.experimental import (
-    read_dispatched,
-    read_elem,
-    write_dispatched,
-    write_elem,
-)
+from anndata.experimental import read_dispatched, write_dispatched
 from anndata.tests.helpers import assert_equal, gen_adata
 
 
@@ -29,7 +24,7 @@ def test_read_dispatched_w_regex():
     adata = gen_adata((1000, 100))
     z = zarr.group()
 
-    write_elem(z, "/", adata)
+    ad.io.write_elem(z, "/", adata)
 
     expected = ad.AnnData(obs=adata.obs, var=adata.var)
     actual = read_dispatched(z, read_only_axis_dfs)
@@ -48,7 +43,7 @@ def test_read_dispatched_dask():
             "awkward-array",
         }:
             # Preventing recursing inside of these types
-            return read_elem(elem)
+            return ad.io.read_elem(elem)
         elif iospec.encoding_type == "array":
             return da.from_zarr(elem)
         else:
@@ -56,7 +51,7 @@ def test_read_dispatched_dask():
 
     adata = gen_adata((1000, 100))
     z = zarr.group()
-    write_elem(z, "/", adata)
+    ad.io.write_elem(z, "/", adata)
 
     dask_adata = read_dispatched(z, read_as_dask_array)
 
@@ -64,7 +59,7 @@ def test_read_dispatched_dask():
     assert isinstance(dask_adata.obsm["array"], da.Array)
     assert isinstance(dask_adata.uns["nested"]["nested_further"]["array"], da.Array)
 
-    expected = read_elem(z)
+    expected = ad.io.read_elem(z)
     actual = dask_adata.to_memory(copy=False)
 
     assert_equal(expected, actual)
@@ -73,10 +68,10 @@ def test_read_dispatched_dask():
 def test_read_dispatched_null_case():
     adata = gen_adata((100, 100))
     z = zarr.group()
-    write_elem(z, "/", adata)
+    ad.io.write_elem(z, "/", adata)
 
-    expected = read_elem(z)
-    actual = read_dispatched(z, lambda _, __, x, **___: read_elem(x))
+    expected = ad.io.read_elem(z)
+    actual = read_dispatched(z, lambda _, __, x, **___: ad.io.read_elem(x))
 
     assert_equal(expected, actual)
 
@@ -101,7 +96,7 @@ def test_write_dispatched_chunks():
         # TODO: Should the passed path be absolute?
         path = "/" + store.path + "/" + k
         if hasattr(elem, "shape") and not isinstance(
-            elem, (sparse.spmatrix, SpArray, ad.AnnData)
+            elem, sparse.spmatrix | SpArray | ad.AnnData
         ):
             if re.match(r"^/((X)|(layers)).*", path):
                 chunks = (M, N)
