@@ -12,7 +12,6 @@ import anndata as ad
 from anndata._io.specs.registry import IORegistryError
 from anndata._io.utils import report_read_key_on_error
 from anndata.compat import _clean_uns
-from anndata.experimental import read_elem, write_elem
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -108,10 +107,10 @@ def test_only_child_key_reported_on_failure(tmp_path, group_fn):
     pattern = r"(?s)^((?!Error raised while writing key '/?a').)*$"
 
     with pytest.raises(IORegistryError, match=pattern):
-        write_elem(group, "/", {"a": {"b": Foo()}})
+        ad.io.write_elem(group, "/", {"a": {"b": Foo()}})
 
-    write_elem(group, "/", {"a": {"b": [1, 2, 3]}})
+    ad.io.write_elem(group, "/", {"a": {"b": [1, 2, 3]}})
     group["a/b"].attrs["encoding-type"] = "not a real encoding type"
 
     with pytest.raises(IORegistryError, match=pattern):
-        read_elem(group)
+        ad.io.read_elem(group)
